Resolve a build failure related to variable naming.
Builds with all warnings enabled failed because there was a variable
named 'contentType' which is not 'snake case', the preferred naming
schema in Rust. While fixing that, it became clear that the variable
was not even needed, as its value was only used for an if-then
conditional.

In addition, 'cargo fmt' was run to ensure the code formatting and
imports are correct.